test(word-index): make per-edit occupancy guard a work count
The per-edit seam's occupancy bound was a wall-clock max-block over a
401-document fixture with retry: 2. It was both load-sensitive and a heavy
neighbour in the timing-sensitive lane: joining that lane forced its cap
from 2 to 1 (#2245 interim) because on a loaded runner the fixture rebuild
starved performance-report-occupancy's sampler.

Assert instead that the seam's cooperative replacement reads the clock a
number of times bounded by the old document's distinct tokens, not by the
posting elements it walks. That is a deterministic work count, invariant to
machine speed and runner load, and it reds two orders of magnitude on the
per-element deadline check the #2067 tax reintroduces. The guard no longer
needs the 401-document fixture or the retry, and it leaves the
timing-sensitive lane, so that lane returns to maxWorkers: 2.

The shared countClockReads helper moves to tests/support/perf-harness.ts,
so the incremental clock-read guard and this seam guard read one
implementation instead of two copies.

Refs #2254
